Move some behaviors into Webpack
