+++ _js/app.js
@@ -6,5 +6,10 @@ window.Prism = require('prismjs')
 require('sanitize.css')
 require('prismjs/plugins/line-highlight/prism-line-highlight.css')
 
+// Behaviors that need to go first
+require('./behaviors/main-body')
+
+// All the others
 function requireAll (r) { r.keys().forEach(r) }
+requireAll(require.context('./initializers/', true, /\.js$/))
 requireAll(require.context('./behaviors/', true, /\.js$/))
