+++ _js/behaviors/main-body.js
@@ -1,0 +1,11 @@
+import $ from 'jquery'
+import wrapify from '../wrapify'
+
+/*
+ * Behavior: Wrapping
+ */
+
+$(function () {
+  const $root = $('[data-js-main-body]')
+  wrapify($root)
+})
