+++ _js/initializers/onmount.js
@@ -1,0 +1,4 @@
+import ready from 'dom101/ready'
+import onmount from 'onmount'
+
+ready(() => { onmount() })
